requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.35.85
botocore==1.35.85
CacheControl==0.14.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
distlib==0.3.9
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-soft-delete==1.0.16
django-softdelete==0.11.4
django-storages==1.14.4
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
firebase-admin==6.6.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
identify==2.6.2
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.2
Markdown==3.7
msgpack==1.1.0
nodeenv==1.9.1
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
phonenumbers==8.13.52
platformdirs==4.3.6
polib==1.2.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
# psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycurl==7.45.4
PyJWT==2.10.1
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
sentry-sdk==2.22.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.27.1
watchfiles==1.0.4
wcwidth==0.2.13
weasyprint==65.1
websockets==15.0.1
